Correction of Interface implementation

+++ src/Calc.java
@@ -1,3 +1,3 @@
 public interface Calc {
-    public int Operate(String input);
+    public float Operate(String input);
 }
